[cli] delete command should drop column tables. Also refactor NodeCoordinator commands.
